Add a cast to fix: tempfile.c(202) : warning C4267: '=' : conversion from 'size_t' to 'unsigned long'
svn path=/trunk/; revision=34911

+++ tempfile.c
@@ -199,7 +199,7 @@ create_tempfile(char **namebuf, const char *pfx)
 	sep[0] = G_DIR_SEPARATOR;
 	tmp_file = g_strconcat(tmp_dir, sep, pfx, "_", timestr, "_", TMP_FILE_SUFFIX, NULL);
 	if (strlen(tmp_file) > tf[idx].len) {
-		tf[idx].len = strlen(tmp_file) + 1;
+		tf[idx].len = (int)strlen(tmp_file) + 1;
 		tf[idx].path = (char *)g_realloc(tf[idx].path, tf[idx].len);
 	}
 	g_strlcpy(tf[idx].path, tmp_file, tf[idx].len);
